fix: split rmcryptpad schema modules
